examples/gemini-live/pyproject.toml
[project]
name = "open-xiaoai-gemini"
version = "1.0.0"
description = "小爱音箱接入 Gemini Live"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "google-genai>=1.12.1",
    "httpx[socks]>=0.28.1",
    "maturin>=1.8.3",
    "numpy>=2.2.3",
]

[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[tool.maturin]
features = ["pyo3/extension-module"]

[tool.uv]
# Rebuild package when any rust files change 
# https://github.com/PyO3/maturin/issues/2314#issuecomment-2488542771
cache-keys = [
    { file = "pyproject.toml" },
    { file = "Cargo.toml" },
    { file = "src/**/*.rs" },
]
